+++ src/event/win_hide.rs
@@ -4,7 +4,7 @@ use rmpv::Value;
 /// Stop displaying the window. The window can be shown again later.
 #[derive(Debug, Clone)]
 pub struct WinHide {
-    pub grid: u64,
+    pub grid: u32,
 }
 
 impl Parse for WinHide {
